// app/enums/ActionTypes.ts
enum ActionTypes {
  SET_APP_CONFIG = 'setAppConfig',
  SHOW_LOADING = 'showLoading',
  DISMISS_LOADING = 'dismissLoading',
  SHOW_DIALOG = 'showDialog',
  DISMISS_DIALOG = 'dismissDialog',
  SHOW_CONFIRM_DIALOG = 'showConfirmDialog',
  SET_CONFIRM_DIALOG_CHOICE = 'setConfirmDialogChoice',
  SHOW_KEYPAD = 'showKeypad',
  HIDE_KEYPAD = 'hideKeypad',
  CLEAR_STORE = 'clearStore',
}

export default ActionTypes;